Wall Wextra Werror

+++ hardware/arduino/zunoG2/libraries/Z-Uno-2G/examples/RF24/MulticeiverDemo/MulticeiverDemo.ino
@@ -190,7 +190,7 @@ void setRole() {
 
     // Set the address on pipe 0 to the RX node.
     radio.stopListening(); // put radio in TX mode
-    radio.openWritingPipe((uint8_t)address[role]);
+    radio.openWritingPipe((uint8_t)address[(uint8_t)role]);
 
     // According to the datasheet, the auto-retry features's delay value should
     // be "skewed" to allow the RX node to receive 1 transmission at a time.
